+++ internal/tui/components/sidebar.go
@@ -15,7 +15,7 @@ import (
 var sidebarLog *log.Logger
 
 func init() {
-	f, err := os.OpenFile(config.LogPath(), os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
+	f, err := os.OpenFile(config.LogPath(), os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0o600)
 	if err != nil {
 		sidebarLog = log.New(os.Stderr, "[sidebar] ", log.Ltime)
 		return