+++ internal/tui/components/statusbar.go
@@ -16,7 +16,7 @@ import (
 var statusLog *log.Logger
 
 func init() {
-	f, err := os.OpenFile(config.LogPath(), os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
+	f, err := os.OpenFile(config.LogPath(), os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0o600)
 	if err != nil {
 		statusLog = log.New(os.Stderr, "[status] ", log.Ltime)
 		return